% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PaleoPopResults.R
\name{PaleoPopResults}
\alias{PaleoPopResults}
\title{R6 class representing paleopop simulator results.}
\description{
\code{\link[R6:R6Class]{R6}} class for encapsulating and dynamically generating
spatially-explicit \code{\link{paleopop_simulator}} results, as well as optional
re-generated \code{\link{Generator}} for niche carrying capacity and/or human
density.
}
\section{Super classes}{
\code{\link[poems:GenericClass]{poems::GenericClass}} -> \code{\link[poems:GenericModel]{poems::GenericModel}} -> \code{\link[poems:SpatialModel]{poems::SpatialModel}} -> \code{\link[poems:SimulationResults]{poems::SimulationResults}} -> \code{PaleoPopResults}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{attached}}{A list of dynamically attached attributes (name-value pairs).}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{model_attributes}}{A vector of model attribute names.}

\item{\code{region}}{A \code{\link{Region}} (or inherited class) object specifying the study region.}

\item{\code{coordinates}}{Data frame (or matrix) of X-Y population coordinates (WGS84) in longitude (degrees West) and latitude (degrees North).}

\item{\code{time_steps}}{Number of simulation time steps.}

\item{\code{burn_in_steps}}{Optional number of initial 'burn-in' time steps to be ignored.}

\item{\code{occupancy_mask}}{Optional binary mask array (matrix), data frame, or raster (stack) for each cell at each time-step of the simulation including burn-in.}

\item{\code{trend_interval}}{Optional time-step range (indices) for trend calculations (assumes indices begin after the burn-in when utilized).}

\item{\code{abundance}}{Matrix of population abundance across simulation time-steps (\emph{populations} rows by \emph{duration} columns).}

\item{\code{abundance_trend}}{Trend or average Sen's \code{\link[trend:sens.slope]{slope}} of total abundance (optionally across a time-step interval).}

\item{\code{ema}}{Matrix of population expected minimum abundance (EMA) across simulation time-steps (\emph{populations} rows by \emph{duration} columns).}

\item{\code{extirpation}}{Array of population extirpation times.}

\item{\code{extinction_location}}{The weighted centroid of cells occupied in the time-step prior to the extirpation of all populations (if occurred).}

\item{\code{harvested}}{Matrix of the number of animals harvested from each population at each time-step (\emph{populations} rows by \emph{duration} columns).}

\item{\code{occupancy}}{Array of the number of populations occupied at each time-step.}

\item{\code{carrying_capacity}}{Optional matrix of simulation input carrying capacity to be combined with results (\emph{populations} rows by \emph{duration} columns).}

\item{\code{human_density}}{Optional matrix of simulation input human density to be combined with results (\emph{populations} rows by \emph{duration} columns).}

\item{\code{all}}{Nested simulation results for all cells.}

\item{\code{parent}}{Parent simulation results for individual cells.}

\item{\code{default}}{Default value/attribute utilized when applying primitive metric functions (e.g. max) to the results.}

\item{\code{attribute_aliases}}{A list of alternative alias names for model attributes (form: \code{alias = "attribute"}) to be used with the set and get attributes methods.}

\item{\code{error_messages}}{A vector of error messages encountered when setting model attributes.}

\item{\code{warning_messages}}{A vector of warning messages encountered when setting model attributes.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-clone}{\code{PaleoPopResults$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="poems" data-topic="GenericModel" data-id="get_attribute">}\href{../../poems/html/GenericModel.html#method-get_attribute}{\code{poems::GenericModel$get_attribute()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="poems" data-topic="GenericModel" data-id="get_attribute_aliases">}\href{../../poems/html/GenericModel.html#method-get_attribute_aliases}{\code{poems::GenericModel$get_attribute_aliases()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="poems" data-topic="SimulationResults" data-id="get_attribute_names">}\href{../../poems/html/SimulationResults.html#method-get_attribute_names}{\code{poems::SimulationResults$get_attribute_names()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="poems" data-topic="SimulationResults" data-id="get_attributes">}\href{../../poems/html/SimulationResults.html#method-get_attributes}{\code{poems::SimulationResults$get_attributes()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="poems" data-topic="SimulationResults" data-id="initialize">}\href{../../poems/html/SimulationResults.html#method-initialize}{\code{poems::SimulationResults$initialize()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="poems" data-topic="SimulationResults" data-id="new_clone">}\href{../../poems/html/SimulationResults.html#method-new_clone}{\code{poems::SimulationResults$new_clone()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="poems" data-topic="SimulationResults" data-id="set_attributes">}\href{../../poems/html/SimulationResults.html#method-set_attributes}{\code{poems::SimulationResults$set_attributes()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PaleoPopResults$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
