% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Region.R
\name{Region}
\alias{Region}
\title{R6 class representing a study region.}
\description{
\code{\link[R6:R6Class]{R6}} class representing a study region of spatial grid cells
defined via a list of longitude/latitude cell coordinates (WGS84), or a
\emph{RasterLayer} object (see \code{\link[raster:raster-package]{raster}}).
}
\section{Super class}{
\code{\link[poems:GenericClass]{poems::GenericClass}} -> \code{Region}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{attached}}{A list of dynamically attached attributes (name-value pairs).}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{coordinates}}{Data frame (or matrix) of X-Y population (WGS84) coordinates in longitude (degrees West) and latitude (degrees North) (get and set), or distance-based coordinates dynamically returned by region raster (get only).}

\item{\code{region_raster}}{A \emph{RasterLayer} object (see \code{\link[raster:raster-package]{raster}}) defining the region with finite values (NAs elsewhere).}

\item{\code{use_raster}}{Boolean to indicate that a raster is to be used to define the region (default TRUE).}

\item{\code{region_cells}}{Dynamically calculated number of region coordinates or raster cells with finite/non-NA values.}

\item{\code{region_indices}}{Dynamically calculated region indices for raster cells with finite/non-NA values (all if not a raster).}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Region$new()}}
\item \href{#method-raster_is_consistent}{\code{Region$raster_is_consistent()}}
\item \href{#method-clone}{\code{Region$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="poems" data-topic="GenericClass" data-id="new_clone">}\href{../../poems/html/GenericClass.html#method-new_clone}{\code{poems::GenericClass$new_clone()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initialization method sets coordinates or raster for region.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Region$new(
  coordinates = NULL,
  template_raster = NULL,
  region_raster = NULL,
  use_raster = TRUE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{coordinates}}{Data frame (or matrix) of X-Y coordinates (WGS84) in longitude (degrees West) and latitude (degrees North).}

\item{\code{template_raster}}{A \emph{RasterLayer} object (see \code{\link[raster:raster-package]{raster}}) defining the region with example finite values (NAs elsewhere)}

\item{\code{region_raster}}{A \emph{RasterLayer} object (see \code{\link[raster:raster-package]{raster}}) defining the region with finite cell indices (NAs elsewhere).}

\item{\code{use_raster}}{Boolean to indicate that a raster is to be used to define the region (default TRUE).}

\item{\code{...}}{Additional parameters passed individually.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-raster_is_consistent"></a>}}
\if{latex}{\out{\hypertarget{method-raster_is_consistent}{}}}
\subsection{Method \code{raster_is_consistent()}}{
Returns a boolean to indicate if a raster is consistent with the region raster (matching extent, resolution, and finite/NA cells).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Region$raster_is_consistent(check_raster)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{check_raster}}{A \emph{RasterLayer}, \emph{RasterStack} or \emph{RasterBrick} object (see \code{\link[raster:raster-package]{raster}}) to check for consistency with the region raster.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Boolean to indicate if the raster is consistent with the region raster.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Region$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
