% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimulatorReference.R
\name{SimulatorReference}
\alias{SimulatorReference}
\title{R6 class for a simulator reference}
\description{
\code{\link[R6:R6Class]{R6}} class for dynamically attaching simulator attributes
and results (passed by reference).
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{attached}}{A list of dynamically attached simulator attributes (name-value pairs).}

\item{\code{results}}{A list of dynamically accessed simulator results (name-value pairs).}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SimulatorReference-clone}{\code{SimulatorReference$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SimulatorReference-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SimulatorReference-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SimulatorReference$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
