% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tasmania_ibra_raster}
\alias{tasmania_ibra_raster}
\title{Thylacine vignette Tasmania IBRA raster}
\format{
A \emph{raster::RasterLayer} object:
\describe{
  \item{dimensions}{32 rows by 40 columns grid}
  \item{resolution}{0.1 by 0.1 degree grid cells}
  \item{extent}{longitude 144.5 to 148.5 degrees; latitude -43.8025 to -40.6025 degrees}
  \item{CRS}{WGS84 longitude-latitude}
  \item{values}{IBRA bioregions defined by cells with values 1 to 9 (as per \code{index} in \code{\link{tasmania_ibra_data}})}
}
}
\source{
TBA
}
\usage{
tasmania_ibra_raster
}
\description{
A \emph{raster} dataset defining the grid cells of the nine Interim
Bioregionalisation of Australia (IBRA) bioregions for the Tasmanian study region of
the Thylacine example vignette.
}
\keyword{datasets}
