% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tasmania_raster}
\alias{tasmania_raster}
\title{Thylacine vignette Tasmania raster}
\format{
A \emph{raster::RasterLayer} object:
\describe{
  \item{dimensions}{32 rows by 40 columns grid}
  \item{resolution}{0.1 by 0.1 degree grid cells}
  \item{extent}{longitude 144.5 to 148.5 degrees; latitude -43.8025 to -40.6025 degrees}
  \item{CRS}{WGS84 longitude-latitude}
  \item{values}{region defined by 795 cells with value of 1 (surrounded by non-region \code{NA} values)}
}
}
\source{
TBA
}
\usage{
tasmania_raster
}
\description{
A \emph{raster} dataset defining the grid cells of the Tasmanian study region for the
Thylacine example vignette.
}
\keyword{datasets}
