% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{thylacine_bounty_record}
\alias{thylacine_bounty_record}
\title{Thylacine vignette bounty record}
\format{
A data frame with 22 rows and 11 variables:
\describe{
  \item{Year}{Year during bounty period from 1888 to 1909}
  \item{Total}{Total Tasmania-wide bounty submitted}
  \item{emph{IBRA columns}}{Bounty submitted for each bioregion (as per \code{abbr} in \code{\link{tasmania_ibra_data}})}
}
}
\source{
TBA
}
\usage{
thylacine_bounty_record
}
\description{
A dataset containing the historical record of the Thylacine bounty numbers submitted
across the Tasmanian study region, and for each of the nine Interim Bioregionalisation
of Australia (IBRA) bioregions for Thylacine example vignette.
}
\keyword{datasets}
