% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{thylacine_example_matrices_rerun}
\alias{thylacine_example_matrices_rerun}
\title{Thylacine vignette demonstration example (re-run) matrices}
\format{
A list containing the following matrices:
\describe{
  \item{bounty_slope}{2,000 row by 3 column matrix of calculated slope of total bounty submitted across 3 intervals for each sample simulation}
  \item{ibra_extirpation}{2,000 row by 9 column matrix of extirpation dates for each IBRA bioregion for each sample simulation}
}
}
\source{
Precalculated demonstration via example simulation re-runs.
}
\usage{
thylacine_example_matrices_rerun
}
\description{
A dataset containing precalculated (re-run) summary matrices for use when running the
Thylacine example vignette in demonstration mode. The values were obtained by running
the vignette code for 10 replicate re-runs of the selected 'best' 200 model
simulations with \code{DEMONSTRATION = FALSE}.
}
\keyword{datasets}
