% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{thylacine_example_metrics}
\alias{thylacine_example_metrics}
\title{Thylacine vignette demonstration example metrics}
\format{
A data frame with 20,000 rows and 4 variables:
\describe{
\item{index}{Example simulation number from 1 to 20,000}
\item{bounty_slope_error}{Root mean squared error (RMSE) from estimated total bounty submitted across three intervals (see vignette)}
\item{ibra_extirpation_error}{RMSE from estimated extirpation date for each IBRA bioregion (see vignette)}
\item{total_extinction}{Total extinction date for each example simulation (\code{NA} when persistent beyond 1967)}
}
}
\source{
Precalculated demonstration via example simulation runs.
}
\description{
A dataset containing precalculated summary metrics for use when running the Thylacine
example vignette in demonstration mode. The values were obtained by running the
vignette code for 20,000 model simulations with \code{DEMONSTRATION = FALSE}.
}
\examples{
data(thylacine_example_metrics)
hist(thylacine_example_metrics$bounty_slope_error)
hist(thylacine_example_metrics$ibra_extirpation_error)
hist(thylacine_example_metrics$total_extinction)

}
