% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{thylacine_example_metrics_rerun}
\alias{thylacine_example_metrics_rerun}
\title{Thylacine vignette demonstration example (re-run) metrics}
\format{
A data frame with 2,000 rows and 4 variables:
\describe{
  \item{index}{Example simulation number from 1 to 2,000}
  \item{bounty_slope_error}{Root mean squared error (RMSE) from estimated total bounty submitted across three intervals (see vignette)}
  \item{ibra_extirpation_error}{RMSE from estimated extirpation date for each IBRA bioregion (see vignette)}
  \item{total_extinction}{Total extinction date for each example simulation (\code{NA} when persistent beyond 1967)}
}
}
\source{
Precalculated demonstration via example simulation re-runs.
}
\usage{
thylacine_example_metrics_rerun
}
\description{
A dataset containing precalculated (re-run) summary metrics for use when running the
Thylacine example vignette in demonstration mode. The values were obtained by running
the vignette code for 10 replicate re-runs of the selected 'best' 200 model
simulations with \code{DEMONSTRATION = FALSE}.
}
\keyword{datasets}
