% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{thylacine_hs_raster}
\alias{thylacine_hs_raster}
\title{Thylacine vignette habitat suitability raster}
\format{
A \emph{raster::RasterLayer} object:
\describe{
  \item{dimensions}{32 rows by 40 columns grid}
  \item{resolution}{0.1 by 0.1 degree grid cells}
  \item{extent}{longitude 144.5 to 148.5 degrees; latitude -43.8025 to -40.6025 degrees}
  \item{CRS}{WGS84 longitude-latitude}
  \item{values}{Estimated habitat suitability values of 0 to 1}
}
}
\source{
TBA
}
\usage{
thylacine_hs_raster
}
\description{
A \emph{raster} dataset defining estimated habitat suitability values for each grid
cells of the Tasmanian study region of the Thylacine example vignette.
}
\keyword{datasets}
