% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tasmania_ibra_data}
\alias{tasmania_ibra_data}
\title{Thylacine vignette Tasmania IBRA data}
\format{
A data frame with 9 rows and 4 variables:
\describe{
  \item{index}{Cross-reference index for each bioregion}
  \item{key}{Additional alphabetical cross-reference for each bioregion}
  \item{abbr}{Abbreviated name for each bioregion}
  \item{name}{Full name for each bioregion}
}
}
\source{
TBA
}
\usage{
tasmania_ibra_data
}
\description{
A dataset describing the nine Interim Bioregionalisation of Australia (IBRA)
bioregions for the Tasmanian study region of the Thylacine example vignette.
}
\keyword{datasets}
